src/pyproject.toml
[project]
name = "vitaelangx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "pdfminer-six>=20250506",
    "pillow>=11.2.1",
    "pymysql>=1.1.1",
    "tk>=0.1.0",
]
